Added shebang line
I have BESST installed as a module¹ on my system and the bin folder is added to the path. So with the shebang line on top i can call the script from anywhere without giving the exact path. To my knowledge it should not do any harm in any other setting.

¹: http://en.wikipedia.org/wiki/Environment_Modules_%28software%29